RDM-8928: Refactor befta step class to move redundant methods
