+++ src/aat/java/uk/gov/hmcts/ccd/datastore/befta/DataStoreTestAutomationAdapter.java
@@ -1,5 +1,7 @@
 package uk.gov.hmcts.ccd.datastore.befta;
 
+import io.cucumber.java.Before;
+import org.junit.AssumptionViolatedException;
 import uk.gov.hmcts.befta.DefaultTestAutomationAdapter;
 import uk.gov.hmcts.befta.dse.ccd.TestDataLoaderToDefinitionStore;
 import uk.gov.hmcts.befta.exception.FunctionalTestException;
@@ -7,11 +9,20 @@ import uk.gov.hmcts.befta.player.BackEndFunctionalTestScenarioContext;
 import uk.gov.hmcts.befta.util.ReflectionUtils;
 import uk.gov.hmcts.ccd.datastore.tests.helper.elastic.ElasticsearchTestDataLoaderExtension;
 
+import java.util.UUID;
+
 import static java.util.Optional.ofNullable;
 
 public class DataStoreTestAutomationAdapter extends DefaultTestAutomationAdapter {
 
     private TestDataLoaderToDefinitionStore loader = new TestDataLoaderToDefinitionStore(this);
+
+    @Before("@elasticsearch")
+    public void skipElasticSearchTestsIfNotEnabled() {
+        if (!ofNullable(System.getenv("ELASTIC_SEARCH_ENABLED")).map(Boolean::valueOf).orElse(false)) {
+            throw new AssumptionViolatedException("Elastic Search not Enabled");
+        }
+    }
 
     @Override
     public void doLoadTestData() {
@@ -41,7 +52,8 @@ public class DataStoreTestAutomationAdapter extends DefaultTestAutomationAdapter
                 throw new FunctionalTestException("Problem getting case id as long", e);
             }
         } else if (key.toString().equals("UniqueString")) {
-            return ScenarioData.getUniqueString();
+            String uniqueID = UUID.randomUUID().toString();
+            return "string-" + uniqueID;
         }
         return super.calculateCustomValue(scenarioContext, key);
     }
